% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bentoRarefy.R
\name{bentoRarefy}
\alias{bentoRarefy}
\title{Filter out OTUs by abundance}
\usage{
bentoRarefy(bentoBox, subsampleSize = "min")
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{abundance}{abundance (default = 0.0)}
}
\description{
This function will remove OTUs below the total abundance of x as a fraction, i.e. 0.5 = 50\%
See http://www.ncbi.nlm.nih.gov/pmc/articles/PMC3531572/ if interested in finding out which threshold to use
}
\examples{
aNewBentoBox = filterByAbundance(bentoBox, column, prevalence = 0.0)
}

