% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOtuBentoBoxFromFiles.R
\name{createOtuBentoBoxFromFiles}
\alias{createOtuBentoBoxFromFiles}
\title{createOtuBentoBoxFromFiles}
\usage{
createOtuBentoBoxFromFiles(f_otutable, f_metadata)
}
\arguments{
\item{dataFile_otutable}{Otu table file}

\item{dataFile_metadata}{Meta data file}
}
\description{
This function creates an OtuBentoBox from two file inputs: Otu table and a Meta data.
}
\examples{
aNewBentoBox = createOtuBentoBoxFromFiles(dataFile_otutable, dataFile_metadata)
}

