% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBasedOnMetadataFromBentoBox.R
\name{filterBasedOnMetadataFromBentoBox}
\alias{filterBasedOnMetadataFromBentoBox}
\title{Filter entries based on user selected metadata}
\usage{
filterBasedOnMetadataFromBentoBox(bentoBox, column, termsToRetain,
  type = "remove")
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{column}{Column in metadata}

\item{termsToRetain}{Terms to retain}

\item{type}{Remove or retain samples. Optinos: "remove", "remain" (default = "remove")}
}
\description{
This function filters entries based on user selected metadata columns from an OTU Bento Box
}
\examples{
aNewBentoBox = filterBasedOnMetadataFromBentoBox(aBentoBox, "Habitat", c("Water", "Water_M", "Biofilm", "Sediment_T"), type = "remove")
}

