% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByReplicatePrevalence.R
\name{filterByReplicatePrevalence}
\alias{filterByReplicatePrevalence}
\title{Filter out OTUs with less than x% prevalence}
\usage{
filterByReplicatePrevalence(bentoBox, column, prevalence = 0.75)
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{column}{Column in metadata to group replicates}

\item{prevalence}{prevalence. So, 0.75 means an OTU which is present in less than 75% of all replicates will be removed.}
}
\description{
Say you have collected 10 replicates. Only 5 out of 10 of the replicates have an OTU. This function filters out such OTUs. The idea is to have a dataset which is a good representative of the site.
}
\examples{
aNewBentoBox =
}

