% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOTUsFromBentoBox.R
\name{filterOTUsFromBentoBox}
\alias{filterOTUsFromBentoBox}
\title{Filter samples from an OTU Bento Box by ID}
\usage{
filterOTUsFromBentoBox(bentoBox, OTUsToRemove, type = "remove")
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{type}{Remove or retain samples. Optinos: "remove", "remain" (default = "remove")}

\item{samplesToRemove}{a vector of samples to remove from Your Bento Box}
}
\description{
This function filter samples from an OTU Bento Box
}
\examples{
aNewBentoBox = filterOTUsFromBentoBox(aBentoBox, c("OTU39359", "OTU2365"), type = "remove")
}

