% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSamplesByIDFromBentoBox.R
\name{filterSamplesByIDFromBentoBox}
\alias{filterSamplesByIDFromBentoBox}
\title{Filter samples from an OTU Bento Box by ID}
\usage{
filterSamplesByIDFromBentoBox(bentoBox, samplesToRemove, type = "remove")
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{samplesToRemove}{a vector of samples to remove from Your Bento Box}

\item{type}{Remove or retain samples. Optinos: "remove", "remain" (default = "remove")}
}
\description{
This function filter samples from an OTU Bento Box
}
\examples{
aNewBentoBox = filterSamplesByIDFromBentoBox(aBentoBox, c("Sample1", "Sample2"), type = "remove")
}

