% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSamplesBySampleSizeFromBentoBox.R
\name{filterSamplesBySampleSizeFromBentoBox}
\alias{filterSamplesBySampleSizeFromBentoBox}
\title{Filter samples from an OTU Bento Box by sample size}
\usage{
filterSamplesBySampleSizeFromBentoBox(bentoBox, sampleSize, type = "remove")
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{sampleSize}{Sample size}

\item{type}{Remove or retain samples. Optinos: "remove", "remain" (default = "remove")}
}
\description{
This function filter samples from an OTU Bento Box below the user defined sample size
}
\examples{
aNewBentoBox = filterSamplesBySampleSizeFromBentoBox(aBentoBox, c("Sample1", "Sample2"), type = "remove")
}

