% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSamplesFromBentoBox.R
\name{filterSamplesFromBentoBox}
\alias{filterSamplesFromBentoBox}
\title{Filter samples from an OTU Bento Box}
\usage{
filterSamplesFromBentoBox(bentoBox, samplesToRemove, type = "remove")
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{type}{Remove or retain samples. Optinos: "remove", "remain" (default = "remove")}

\item{samples}{a vector of samples to remove from Your Bento Box}
}
\description{
This function filter samples from an OTU Bento Box
}
\examples{
aNewBentoBox = filterSamplesFromBentoBox(aBentoBox, c("Sample1", "Sample2"), type = "remove")
}

