% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTaxaFromBentoBox.R
\name{filterTaxaFromBentoBox}
\alias{filterTaxaFromBentoBox}
\title{Filter taxa from an OTU Bento Box}
\usage{
filterTaxaFromBentoBox(bentoBox, classificationLevel, nameToRemove,
  type = "remove")
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{classificationLevel}{Your classificationLevel}

\item{nameToRemove}{Names to remove}

\item{type}{Remove or retain samples. Optinos: "remove", "remain" (default = "remove")}
}
\description{
This function filters taxa from an OTU Bento Box
}
\examples{
aNewBentoBox = filterTaxaFromBentoBox(aBentoBox, "k", "k__Bacteria", type = "remove")
}

