% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleIDs.R
\name{getSampldIDs}
\alias{getSampldIDs}
\title{getSampldIDs}
\usage{
getSampldIDs(otutable_filename)
}
\arguments{
\item{otu_table}{Your OTU Table.}
}
\description{
This function retrieves Sample IDs from the OTU table.
}
\examples{
SampleIDs = getSampldIDs("otu_table.txt")
}
\keyword{OTU}
\keyword{sampleIDs}

