% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleNames.R
\name{getSampleNames}
\alias{getSampleNames}
\title{getSampleNames}
\usage{
getSampleNames(file)
}
\arguments{
\item{file}{the name of the OTU table file which the data are to be read from. The OTU table must have derived from biom.}
}
\description{
This function retrieves sample names from an OTU table.
}
\examples{
sampleNames = getSampleNames("otu_table.txt")
}
\keyword{OTU}
\keyword{names}
\keyword{sample}

