% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otubasket.R
\docType{class}
\name{otubasket-class}
\alias{otubasket}
\alias{otubasket-class}
\title{Class otubasket}
\usage{
\S4method{initialize}{otubasket}(.Object, otutable_filename, metadata_filename)

otubasket(otutable_filename, metadata_filename)
}
\description{
Class \code{otubasket} defines a gas sensor device.

Constructor method of otubasket Class.

Wrapper function Sensor.
}

