% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otuBentoBox.R
\docType{class}
\name{otuBentoBox-class}
\alias{otuBentoBox}
\alias{otuBentoBox-class}
\title{Class otuBentoBox}
\usage{
\S4method{initialize}{otuBentoBox}(.Object, otutable, metadata, taxonomy)

otuBentoBox(otutable, metadata, taxonomy)
}
\description{
Class \code{otuBentoBox} defines an OTU Bento Box class

Constructor method of otuBentoBox Class.

Wrapper function otuBentoBox.
}

