% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBasedOnMetadataFromBentoBox.R, R/filterSamplesFromBentoBox.R
\name{otuBentoBox}
\alias{filterBasedOnMetadataFromBentoBox}
\alias{filterSamplesFromBentoBox}
\alias{otuBentoBox}
\title{Filter entries based on user selected metadata}
\usage{
filterBasedOnMetadataFromBentoBox(bentoBox, column, termsToRetain,
  type = "remove")

filterSamplesFromBentoBox(bentoBox, samplesToRemove, type = "remove")
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{column}{Column in metadata}

\item{termsToRetain}{Terms to retain}

\item{bentoBox}{Your Bento Box}

\item{samples}{Samples to remove from Your Bento Box}

\item{remove}{Whether of not you want to remove or retain (default = TRUE)}
}
\description{
Filter entries based on user selected metadata from an OTU Bento Box

Filter samples from an OTU Bento Box
}
\examples{
aNewBentoBox = filterBasedOnMetadataFromBentoBox(otuBentoBox_16S, "Habitat", c("Water_GFA", "Water_M", "Biofilm", "Sediment_T"))
aNewBentoBox = filterSamplesFromBentoBox(aBentoBox, c("Sample1", "Sample2"))
}

