% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolReplicates.R
\name{poolReplicates}
\alias{poolReplicates}
\title{Pool replicates}
\usage{
poolReplicates(bentoBox, column)
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{column}{Column in metadata to group replicates}
}
\description{
Resulting table is the mean count.
}
\examples{
aNewBentoBox = poolReplicates(bentoBox, "replicateGroup")
}

