% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOtuBentoBoxFromFiles.R
\name{createOtuBentoBoxFromFiles}
\alias{createOtuBentoBoxFromFiles}
\title{createOtuBentoBoxFromFiles}
\usage{
createOtuBentoBoxFromFiles(f_otutable, f_metadata)
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{classificationLevel}{Your classificationLevel}

\item{nameToRemove}{Names to remove}

\item{type}{Remove or retain samples. Optinos: "remove", "remain" (default = "remove")}
}
\description{
This function filters taxa from an OTU Bento Box
}
\examples{
aNewBentoBox = createOtuBentoBoxFromFiles(aBentoBox, "k", "k__Bacteria", type = "remove")
}

