% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByAbundance.R
\name{filterByAbundance}
\alias{filterByAbundance}
\title{Filter out OTUs by abundance}
\usage{
filterByAbundance(bentoBox, abundance = 0.75)
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{abundance}{abundance (default = 0.0)}
}
\description{
This function will remove OTUs below the total abundance of x as a fraction, i.e. 0.5 = 50\%
See http://www.ncbi.nlm.nih.gov/pmc/articles/PMC3531572/ if interested in finding out which threshold to use
}
\examples{
aNewBentoBox = filterByAbundance(bentoBox, abundance = 0.5)
}

