% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByReplicatePrevalence.R
\name{filterByReplicatePrevalence}
\alias{filterByReplicatePrevalence}
\title{Filter out OTUs with less than x\% prevalence}
\usage{
filterByReplicatePrevalence(bentoBox, column, prevalence = 0.75)
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{column}{Column in metadata to group replicates}

\item{prevalence}{prevalence. So, 0.5 means an OTU which is present in less than 50\% of all replicates will be removed.}
}
\description{
Say you have collected 10 replicates.
Only 5 of 10 of the replicates have a particular OTU.
You want to retain such OTUs but remove those with less than 5.
This function will deal with this situation! Choose prevalence = 0.5 to achieve this.
}
\examples{
aNewBentoBox = filterByReplicatePrevalence(bentoBox, column, prevalence = 0.5)
}

