% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTaxaFromBentoBox.R
\name{filterTaxaFromBentoBox}
\alias{filterTaxaFromBentoBox}
\title{Remove taxa from an OTU Bento Box}
\usage{
filterTaxaFromBentoBox(bentoBox, classificationLevel, nameToRemove,
  remove = TRUE)
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{classificationLevel}{Your classificationLevel}

\item{nameToRemove}{Names to remove}

\item{remove}{Whether of not you want to remove or retain (default = TRUE)}
}
\description{
This function removes taxa from an OTU Bento Box
}
\examples{
aNewBentoBox = filterTaxaFromBentoBox(aBentoBox, "k", "k__Bacteria", remove = TRUE)
}

