% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleNames.R
\name{getSampleNames}
\alias{getSampleNames}
\title{getSampleNames}
\usage{
getSampleNames(.f_otu_table)
}
\arguments{
\item{otu_table}{OTU Table file}
}
\description{
This function retrieves sample names from the OTU table. The OTU table must be derived from biom.
}
\examples{
sampleNames = getSampleNames("otu_table.txt")
}
\keyword{OTU}
\keyword{names}
\keyword{sample}

