% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otuBentoBox.R
\docType{class}
\name{otuBentoBox-class}
\alias{otuBentoBox}
\alias{otuBentoBox-class}
\title{Class otuBentoBox}
\usage{
\S4method{initialize}{otuBentoBox}(.Object, f_otutable, f_filename)

otuBentoBox(otutable_filename, metadata_filename)
}
\description{
Class \code{otuBentoBox} defines a gas sensor device.

Constructor method of otuBentoBox Class.

Wrapper function otuBentoBox.
}

