% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otuBentoBox.R
\name{otuBentoBox}
\alias{otuBentoBox}
\alias{removeTaxaFromBentoBox}
\alias{retainTaxaFromBentoBox}
\title{Remove taxa from an OTU Bento Box}
\usage{
removeTaxaFromBentoBox(bentoBox, classificationLevel, nameToRemove)

retainTaxaFromBentoBox(bentoBox, classificationLevel, nameToRemove)
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{classificationLevel}{Your classificationLevel}

\item{nameToRemove}{Names to remove}

\item{bentoBox}{Your Bento Box}

\item{classificationLevel}{Your classificationLevel}

\item{nameToRetain}{Names to remove}
}
\description{
Remove taxa from an OTU Bento Box

Retain taxa from an OTU Bento Box
}
\examples{
aNewBentoBox = removeTaxaFromBentoBox(aBentoBox, "k", "k__Bacteria")
aNewBentoBox = retainTaxaFromBentoBox(aBentoBox, "k", "k__Bacteria")
}

