% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otuBentoBox.R
\name{otuBentoBox}
\alias{otuBentoBox}
\alias{removeSamplesFromBentoBox}
\alias{removeTaxaFromBentoBox}
\alias{retainBasedOnMetadataFromBentoBox}
\alias{retainTaxaFromBentoBox}
\title{Remove taxa from an OTU Bento Box}
\usage{
removeTaxaFromBentoBox(bentoBox, classificationLevel, nameToRemove)

retainTaxaFromBentoBox(bentoBox, classificationLevel, nameToRemove)

removeSamplesFromBentoBox(bentoBox, samplesToRemove)

retainBasedOnMetadataFromBentoBox(bentoBox, column, termsToRetain)
}
\arguments{
\item{bentoBox}{Your Bento Box}

\item{classificationLevel}{Your classificationLevel}

\item{nameToRemove}{Names to remove}

\item{column}{Column in metadata}

\item{termsToRetain}{Terms to retain}

\item{bentoBox}{Your Bento Box}

\item{classificationLevel}{Your classificationLevel}

\item{nameToRetain}{Names to remove}

\item{bentoBox}{Your Bento Box}

\item{samples}{Samples to remove from Your Bento Box}

\item{bentoBox}{Your Bento Box}
}
\description{
Remove taxa from an OTU Bento Box

Retain taxa from an OTU Bento Box

Remove whole samples from an OTU Bento Box

Remove entries based on user selected metadata from an OTU Bento Box
}
\examples{
aNewBentoBox = removeTaxaFromBentoBox(aBentoBox, "k", "k__Bacteria")
aNewBentoBox = retainTaxaFromBentoBox(aBentoBox, "k", "k__Bacteria")
aNewBentoBox = removeSamplesFromBentoBox(aBentoBox, c("Sample1", "Sample2"))
aNewBentoBox = retainBasedOnMetadataFromBentoBox(otuBentoBox_16S, "Habitat", c("Water_GFA", "Water_M", "Biofilm", "Sediment_T"))
}

